refactor: replace InfoDialog with widget-based implementation
- Replace 308-line info_dialog.rs with 92-line widget implementation
- Use DialogWidget for consistent UI with proper control instructions
- Preserve original fallback URL display behavior with proper dialog box
- Keep original keyboard controls: [SPACE] select, [ESC] close, [↑↓/JK] navigate
- Add info_dialog_old.rs for reference
- Reduce code duplication by 216 lines
